+++ src/Presentation/SmartStore.Web.Framework/Security/Captcha/ValidateCaptchaAttribute.cs
@@ -13,9 +13,9 @@ using SmartStore.Utilities;
 
 namespace SmartStore.Web.Framework.Security
 {
-	public class CaptchaValidatorAttribute : ActionFilterAttribute
+	public class ValidateCaptchaAttribute : ActionFilterAttribute
     {
-		public CaptchaValidatorAttribute()
+		public ValidateCaptchaAttribute()
 		{
 			Logger = NullLogger.Instance;
 		}